+++ site/src/pages/blogs.tsx
@@ -1,0 +1,3 @@
+import BlogIndexPage from '@/features/blog/pages'
+
+export default BlogIndexPage
